refactor: simplified some frontend components
