+++ web/src/styles/globals.css
@@ -4,58 +4,51 @@
 
 @layer base {
   :root {
-    --background: 0 0% 100%;
-    --foreground: 222.2 84% 4.9%;
-    --card: 0 0% 100%;
-    --card-foreground: 222.2 84% 4.9%;
-    --popover: 0 0% 100%;
-    --popover-foreground: 222.2 84% 4.9%;
-    --primary: 222.2 47.4% 11.2%;
-    --primary-foreground: 210 40% 98%;
-    --secondary: 210 40% 96.1%;
-    --secondary-foreground: 222.2 47.4% 11.2%;
-    --muted: 210 40% 96.1%;
-    --muted-foreground: 215.4 16.3% 46.9%;
-    --accent: 210 40% 96.1%;
-    --accent-foreground: 222.2 47.4% 11.2%;
-    --destructive: 0 84.2% 60.2%;
-    --destructive-foreground: 210 40% 98%;
-    --border: 214.3 31.8% 91.4%;
-    --input: 214.3 31.8% 91.4%;
-    --ring: 222.2 84% 4.9%;
-    --radius: 0.5rem;
-    --chart-1: 12 76% 61%;
-    --chart-2: 173 58% 39%;
-    --chart-3: 197 37% 24%;
-    --chart-4: 43 74% 66%;
-    --chart-5: 27 87% 67%;
+    /* Light Mode Variables */
+    --background: 0 0% 100%; /* White */
+    --foreground: 0 0% 0%; /* Black */
+
+    --card-background: 0 0% 100%; /* White */
+    --card-foreground: 0 0% 0%; /* Black */
+
+    --muted-background: 0 0% 98%; /* Lighter white */
+    --muted-foreground: 0 0% 60%; /* Gray */
+
+    --primary: 0 0% 0%; /* Black */
+    --primary-foreground: 0 0% 100%; /* White */
+
+    --secondary: 210 100% 50%; /* Blue */
+    --secondary-foreground: 0 0% 100%; /* White */
+
+    --accent: 210 100% 50%; /* Blue */
+    --accent-foreground: 0 0% 100%; /* White */
+
+    --border: 0 0% 96.9%; /* Light gray */
+    --hover: 210 16% 90%; /* Light gray */
+    --input: 210 16% 90%; /* Light gray */
+    --ring: 210 100% 50%; /* Blue */
+
+    --radius: 0.5rem; /* Default border radius */
   }
 
   .dark {
-    --background: 222.2 84% 4.9%;
-    --foreground: 210 40% 98%;
-    --card: 222.2 84% 4.9%;
-    --card-foreground: 210 40% 98%;
-    --popover: 222.2 84% 4.9%;
-    --popover-foreground: 210 40% 98%;
-    --primary: 210 40% 98%;
-    --primary-foreground: 222.2 47.4% 11.2%;
-    --secondary: 217.2 32.6% 17.5%;
-    --secondary-foreground: 210 40% 98%;
-    --muted: 217.2 32.6% 17.5%;
-    --muted-foreground: 215 20.2% 65.1%;
-    --accent: 217.2 32.6% 17.5%;
-    --accent-foreground: 210 40% 98%;
-    --destructive: 0 62.8% 30.6%;
-    --destructive-foreground: 210 40% 98%;
-    --border: 217.2 32.6% 17.5%;
-    --input: 217.2 32.6% 17.5%;
-    --ring: 212.7 26.8% 83.9%;
-    --chart-1: 220 70% 50%;
-    --chart-2: 160 60% 45%;
-    --chart-3: 30 80% 55%;
-    --chart-4: 280 65% 60%;
-    --chart-5: 340 75% 55%;
+    /* Dark Mode Variables */
+    --background: 222.2 84% 5%; /* Black */
+    --foreground: 0 0% 100%; /* White */
+
+    --card-background: 222.2 84% 5%; /* Black */
+    --card-foreground: 0 0% 100%; /* White */
+
+    --muted-background: 222.2 84% 10%; /* Darker black */
+    --muted-foreground: 0 0% 60%; /* Gray */
+
+    --primary: 0 0% 100%; /* White */
+    --primary-foreground: 0 0% 0%; /* Black */
+
+    --border: 210 16% 20%; /* Darker gray */
+    --hover: 210 16% 20%; /* Darker gray */
+    --input: 210 16% 20%; /* Darker gray */
+    --ring: 210 100% 50%; /* Blue */
   }
 }
 
@@ -66,5 +59,6 @@
 
   body {
     @apply bg-background text-foreground;
+    color: var(--foreground);
   }
 }
